# docs/api/openapi/mesas.openapi.yaml
...
openapi: "3.0.3"
info:
  title: "Mesas API — Artifício RPG"
  description: "API do módulo Mesas. Metadados x-artificio-* definidos por heurística de path."
  version: "0.1.0"
servers:
  - url: https://mesas.artificiorpg.com
    description: Produção
  - url: http://localhost:4000
    description: Desenvolvimento
security: []
paths:
...
  /api/v1/admin/scenario-suggestions:
    get:
      operationId: getapi_v1_admin_scenario_suggestions
      summary: "Consulta api v1 admin scenario-suggestions"
      x-artificio-owner: mesas
      x-artificio-scope: admin
      x-artificio-status: active
      x-artificio-auth: admin
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...
  /api/v1/scenario-suggestions:
    post:
      operationId: postapi_v1_scenario_suggestions
      summary: "Cria ou executa api v1 scenario-suggestions"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/scenario-suggestions/mine:
    get:
      operationId: getapi_v1_scenario_suggestions_mine
      summary: "Consulta api v1 scenario-suggestions mine"
      x-artificio-owner: mesas
      x-artificio-scope: self-service
      x-artificio-status: active
      x-artificio-auth: user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/scenarios:
    get:
      operationId: getapi_v1_scenarios
      summary: "Consulta api v1 scenarios"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: none
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/scenarios/{id}:
    get:
      operationId: getapi_v1_scenarios_id
      summary: "Consulta api v1 scenarios id"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: none
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...
  /api/v1/upload:
    post:
      operationId: postapi_v1_upload
      summary: "Cria ou executa api v1 upload"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/upload/url:
    post:
      operationId: postapi_v1_upload_url
      summary: "Cria ou executa api v1 upload url"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...
  /api/v1/vtt-platforms/suggest:
    post:
      operationId: postapi_v1_vtt_platforms_suggest
      summary: "Cria ou executa api v1 vtt-platforms suggest"
      x-artificio-owner: mesas
      x-artificio-scope: self-service
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...